requirements.txt
flask
python-dotenv
requests
fuzzywuzzy
nltk
mysql-connector-python
python-dateutil